+++ flutter_inappwebview_forge_platform_interface/lib/src/context_menu/context_menu.dart
@@ -16,6 +16,7 @@ part 'context_menu.g.dart';
           'To make it work properly on Android, JavaScript should be enabled!',
     ),
     IOSPlatform(),
+    MacOSPlatform(),
   ],
 )
 @ExchangeableObject()